game-flow: move GF_Shutdown to libtrx
