+++ src/tr2/game/game_flow/common.c
@@ -7,53 +7,6 @@
 #include <libtrx/log.h>
 #include <libtrx/memory.h>
 
-static void M_FreeSequence(GAME_FLOW_SEQUENCE *sequence);
-static void M_FreeInjections(INJECTION_DATA *injections);
-static void M_FreeLevel(GAME_FLOW_LEVEL *level);
-static void M_FreeLevels(GAME_FLOW_LEVEL **levels, int32_t *level_count);
-static void M_FreeFMVs(GAME_FLOW *gf);
-
-static void M_FreeSequence(GAME_FLOW_SEQUENCE *const sequence)
-{
-    Memory_Free(sequence->events);
-}
-
-static void M_FreeInjections(INJECTION_DATA *const injections)
-{
-    for (int32_t i = 0; i < injections->count; i++) {
-        Memory_FreePointer(&injections->data_paths[i]);
-    }
-    Memory_FreePointer(&injections->data_paths);
-}
-
-static void M_FreeLevel(GAME_FLOW_LEVEL *const level)
-{
-    Memory_FreePointer(&level->path);
-    Memory_FreePointer(&level->title);
-    M_FreeInjections(&level->injections);
-    M_FreeSequence(&level->sequence);
-}
-
-static void M_FreeLevels(GAME_FLOW_LEVEL **levels, int32_t *const level_count)
-{
-    if (levels != NULL) {
-        for (int32_t i = 0; i < *level_count; i++) {
-            M_FreeLevel(&(*levels)[i]);
-        }
-        Memory_FreePointer(levels);
-    }
-    *level_count = 0;
-}
-
-static void M_FreeFMVs(GAME_FLOW *const gf)
-{
-    for (int32_t i = 0; i < gf->fmv_count; i++) {
-        Memory_FreePointer(&gf->fmvs[i].path);
-    }
-    Memory_FreePointer(&gf->fmvs);
-    gf->fmv_count = 0;
-}
-
 int32_t GF_GetLevelCount(const GAME_FLOW_LEVEL_TYPE level_type)
 {
     switch (level_type) {
@@ -87,26 +40,6 @@ void GF_SetLevelTitle(GAME_FLOW_LEVEL *const level, const char *const title)
 int32_t GF_GetGymLevelNum(void)
 {
     return g_GameFlow.gym_enabled ? LV_GYM : -1;
-}
-
-void GF_Shutdown(void)
-{
-    GAME_FLOW *const gf = &g_GameFlow;
-
-    for (int32_t i = 0; i < gf->injections.count; i++) {
-        Memory_FreePointer(&gf->injections.data_paths[i]);
-    }
-    Memory_FreePointer(&gf->injections.data_paths);
-
-    M_FreeLevels(&gf->levels, &gf->level_count);
-    M_FreeLevels(&gf->demos, &gf->demo_count);
-    M_FreeLevels(&gf->cutscenes, &gf->cutscene_count);
-    M_FreeFMVs(gf);
-
-    if (gf->title_level != NULL) {
-        M_FreeLevel(gf->title_level);
-        Memory_FreePointer(&gf->title_level);
-    }
 }
 
 GAME_FLOW_LEVEL *GF_GetLevel(
